Keep provider panel tests within source-size policy
